TASK: Move all fakes to `Neos\ContentRepository\TestSuite\Fakes`
