+++ Neos.ContentRepository.TestSuite/Classes/Fakes/FakeClock.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Neos\ContentRepository\TestSuite\Behavior\Features\Bootstrap\Helpers;
+namespace Neos\ContentRepository\TestSuite\Fakes;
 
 use DateTimeImmutable;
 use Psr\Clock\ClockInterface;
